Time: 29 ms (19.97%), Space: 17.4 MB (83.7%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -1,18 +1,7 @@
-/**
- * Definition for a binary tree node.
- * struct TreeNode {
- *     int val;
- *     TreeNode *left;
- *     TreeNode *right;
- *     TreeNode() : val(0), left(nullptr), right(nullptr) {}
- *     TreeNode(int x) : val(x), left(nullptr), right(nullptr) {}
- *     TreeNode(int x, TreeNode *left, TreeNode *right) : val(x), left(left), right(right) {}
- * };
- */
 class Solution {
 public:
 
-    int dfs(TreeNode* root, int targetSum) {
+    int dfs(TreeNode* root, long long targetSum) {
         if (!root) return 0;
         return (root->val == targetSum) + dfs(root->left, targetSum - root->val) + dfs(root->right, targetSum - root->val);
     }
